Redesign the trophy display and add a trophies component.
This introduces `.rb-c-trophies`, a formal component for showing
trophies. This replaces the old `.box.trophies`, and is quite different.
It's the first big visual update for trophies since 1.0.

The new styling doesn't enclose trophies in a box, and isn't
left-aligned. Instead, these are now shown outside of a box, centered on
the page, with a trophy on either side of the text. The trophies are
slightly slanted away from the text.

The text is now a golden-ish color (which is a bit brighter on dark
mode, and a darker gold on light mode).

Removing the banner and adding the slanted trophies gives this a little
bit more personality. In the future, new trophy images could really help
make this stand out.

Testing Done:
Tested on light and dark modes on mobile and desktop, with both sparkly
and fish trophies.

Reviewed at https://reviews.reviewboard.org/r/13746/