Fix menu styling in the top bar.
With recent changes, the menu styling in the top bar became the default
Ink menu color. We now customize the styling here, and also set the
menus to align to the right edge of the dropdown handles.

Text color for the headers was also incorrect for the color being used
when on light mode. We now set that to the intended value as well.

Testing Done:
Tested the top bar and its menus in light and dark modes. They now look
as they did before.

Reviewed at https://reviews.reviewboard.org/r/13757/