Fix screenshot review UI comment de/serialization.
The comment serialization for the screenshot review UI was a bit broken,
trying to pull the wrong keys out of the serialized comment dict.

This change fixes that, and updates it so we're using our standard
region comment serialization types for everything.

Testing Done:
Created a review request with a screenshot via the admin DB forms. Added
a comment to that screenshot and published the review. Reopened the
review UI and saw that the comment was properly loaded back into the
review UI without any errors.

Reviewed at https://reviews.reviewboard.org/r/13768/