Bullet-proof {% avatar %} templatetag rendering.
We were getting an error email from a bot indexing the search page that
was attempting to render the avatar when the user is `None`. This change
adds some bullet-proofing so we don't hit an error in that case.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13769/