Fix commits list with draft diffs.
When viewing a draft diff, the commits list in the diff viewer would not
render properly. This was happening because the query for all commits
was not including the draft diffset. This change fixes that up and adds
unit tests for the two cases of published diff and draft diff.

Testing Done:
- Verified that I could view the commits list for a draft diff.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13770/