Fix draft comments on unpublished diffs.
We weren't including the draft's diffset in the list of all diffsets,
which could cause a KeyError when attempting to compute the comments
hint data when there was a comment on an unpublished draft. This change
fixes that so we include the diffset and all its filediffs in the list
of all diffsets and filediffs.

Testing Done:
Reproduced the bug with comments on draft diffs and verified that it no
longer occurred.

Reviewed at https://reviews.reviewboard.org/r/13777/