Add DiffSet to clean-orphaned-data.
Apparently DiffSet is another type of object that we can leave orphaned
in certain situations. This change adds it to the clean-orphaned-data
command.

While I was here I cleaned things up a bit to remove unused parameters
and plumb through a missing argument that was causing us to count things
twice.

Testing Done:
Created a bunch of orphaned DiffSet objects and ran this to delete them.

Reviewed at https://reviews.reviewboard.org/r/13787/