Fix num_diffs in diffviewer context.
When I added the draft diffset to the diffsets list that's included in
the diffviewer page context, I missed that we were also adding 1 to the
num_diffs when there was a draft diffset. This was causing the diff
revision slider to show an extra revision that didn't exist.

Testing Done:
Posted a change and saw that my draft said "Diff Revision 1 (Latest)"
instead of showing me a slider with two revisions.

Reviewed at https://reviews.reviewboard.org/r/13791/