Update the diff file index for dark mode, and fix rendering issues.
The diff file index now displays with a darker palette on dark mode,
fitting in with the rest of the unified banner and the review request
box.

The styling is based on the Ink data table styling, with uses a black
background, helping the graphs really pop.

While working on this, I noticed that the dividing lines between the
rows were blurred slightly in light mode. This turned out to be due to
the scrolling placement (via `translateY`) and box height, which was
using sub-pixel positioning. By rounding both of these value, we align
everything to the pixel grid, keeping presentation sharp.

Icon and text alignment has also been fixed. This stood out a lot more
on dark mode.

Testing Done:
Tested in both light and dark modes, docked and undocked.

Reviewed at https://reviews.reviewboard.org/r/13798/