Fix up a bunch of typos and misspellings.
This change fixes up the vast majority of typos and misspellings in our
codebase. The one non-false-positive hit that I've chosen not to change
is the "collapsable" key in the diff chunks, since that data is stored
in cache and might cause bad behavior after an upgrade.

Testing Done:
- Ran python tests.
- Ran js tests.

Reviewed at https://reviews.reviewboard.org/r/13806/