Update Review Board datagrids for dark mode.
This updates our datagrids to remove a bunch of CSS overrides that are
now present in Djblets, and to convert what's left on the pages to dark
mode.

Styling updates include the content header, sidebar, custom columns, and
user page.

The user page has had a tweak to the tabs. We now use `rb-c-tabs`
instead of a custom tab implementation, ensuring we use the same style
consistently. This is primarily a styling update. In the future, this
will move to an Ink component, once that's written.

There's also a fix for the Groups page, which was using the default
template and not our custom template. This was made apparent with the
removal of some of our custom styling.

Testing Done:
Unit tests pass.

Tested all the datagrid pages in light and dark mode.

See https://reviews.reviewboard.org/r/13809/ for screenshots.

Reviewed at https://reviews.reviewboard.org/r/13810/