Disable dark mode for the admin UI.
Due to time constraints, dark mode for the administration UI is out of
scope for Review Board 7. For now, we're disabling that entirely there,
using light mode only.

This is done by wrapping the theme-related attributes in a block and
disabling it in the admin UI template.

We're also setting the header background color for the admin UI to get
the standard yellow bar back.

Testing Done:
Ran through the admin UI with dark mode on. It looked as it should in a
standard light mode.

Reviewed at https://reviews.reviewboard.org/r/13820/