Fix dark mode styling for diff comment tooltips.
When hovering over the diff comments on the left of the diff viewer, the
comment preview was showing white on green or white on blue, instead of
black. Same with the save notification bubble when saving or deleting a
comment. This was due to inheriting the default text color on dark mode.

We now force the text color to be an on-light color.

Testing Done:
Tested comment previews and save notifications in both light and dark
mode.

Reviewed at https://reviews.reviewboard.org/r/13822/