Fix the color of field section labels for diffs in dark mode.
The Diff Revision, Commits, and Files section headers in the diff viewer
page had a dark grey background, which conflicted with the background
and didn't match the styling of the rest of the page.

We unfortunately don't have any standard classes for these that we can
depend on. This should be addressed, but not for this release, as it'd
introduce too many structural changes that aren't worth fixing up right
now.

In the meantime, we're now referencing the field label styling via CSS
variables. We're also using the proper font size for section labels,
which is smaller than we had before, but more consistent with the rest
of the page.

Testing Done:
Tested the review request and diff viewer pages in light and dark mode,
and compared the styling to the other section headers.

Reviewed at https://reviews.reviewboard.org/r/13823/