Fix the updates bubble on dark mode and center the bubble.
The updates bubble has been updated to work correctly on dark mode. We
still use the same background color (utilizing the Ink warning alert
background color), but the text and link colors are fixed to look
correct on it.

The bubble is now centered horizontally, instead of off to the side.
This mirrors the design we've had in RBCommons for years, and helps keep
the bubble better in view on wide screens.

Testing Done:
Tested on light and dark modes.

Reviewed at https://reviews.reviewboard.org/r/13829/