Fix "Update Diff" to do the right thing on review requests with history.
When a review request has commit history, it needs to always be updated
via RBTools. Attempting to use "Update Diff" to upload a diff file will
break in a confusing and useless way.

This change makes it so we show an informative error instead of letting
the user choose a diff file and then showing them an unknown error.

This also gets rid of an obsolete implementation for this that hadn't
been removed from ReviewRequestEditorView.

Testing Done:
Clicked "Update Diff" on a review request with commit history and saw
the new error dialog. Did it on a review request without commit history
and saw the upload dialog as before.

Reviewed at https://reviews.reviewboard.org/r/13846/