Add dark mode for infoboxes.
Infoboxes now work in dark mode, using a typical dark grey background
and a light grey contrasting outline.

This applies to all the variations of infoboxes. That involved fixing
fade-out gradients on the review request and bug infoboxes, and the Ship
It bar and issue status bar on the review request infoboxes.

The Ship It and issue status areas don't use the same visual flair as on
light mode, instead relegating the responsibility to the icon prefixes
(of which there can be multiple for a given review request anyway).

Testing Done:
Tested all the infoboxes with all visual states in light and dark modes.

Reviewed at https://reviews.reviewboard.org/r/13824/