Convert RB.HeaderView to TypeScript/spina.
This change converts our HeaderView to TypeScript. This also fixes a
couple lost-standing annoying issues with our unit tests where some of
the page views being tested were inadvertently connecting to the page's
global HeaderView and then removing it when the test was complete. There
was also one admin-specific style that was leaking into the test page's
display of the header, so I've changed that to be specific to admin
pages.

Reviewed at https://reviews.reviewboard.org/r/14061/