Fix the CodeMirror cursor display in dark mode.
We were inheriting the default style for the CodeMirror cursor in some
uses, which resulted in a black-on-black cursor. We were overriding this
correctly in rich text fields, but not elsewhere.

This change moves the fixed styling into the common CodeMirror
definition, fixing that. While I was at it, I also removed the
now-redundant selection styling, utilizing the common definition that
was placed there before.

Testing Done:
Tested the CodeMirror usage in both review request fields and in the
API Token Policy editor, checking both the cursor and the selections to
make sure they were visible in light and dark modes.

Reviewed at https://reviews.reviewboard.org/r/13876/