Fix overlapping elements in mobile mode for the datagrid filter.
When viewing a datagrid in mobile mode, we were absolute-positioning the
filter dropdown to the right. This used to be fine, since we were
restricting the size of the title text, but the recent move to flexbox
changed that, resulting in the elements overlapping.

Fortunately, we can just take advantage of that flexbox and get rid of
the absolute positioning.

Testing Done:
Tested the datagrid filter display at varying widths, making sure that
wrapping was always correct.

Reviewed at https://reviews.reviewboard.org/r/13875/