Fix help text and API token policy menu link colors.
This updates the help text to be visible in dark mode. We now use the
weak foreground color in Ink for this text, ensuring it'll look fine in
both light and dark modes.

The API token policy menu's "Custom" menu item was hard-coded to blue to
resemble a link. We now use the Ink link menu foreground color instead.

Testing Done:
Tested the menu and the link in light and dark modes.

Made sure it was clickable and visible.

Reviewed at https://reviews.reviewboard.org/r/13858/