Fix the display of diff errors in dark mode.
Diff errors had a white-on-yellow background on dark mode, due to using
a hard-coded background color and the default text color.

We now use the text-on-light color, and the warning alert color for the
background. This matches the warning alert style we use elsewhere, which
is designed to be quite noticeable on light and dark modes.

Testing Done:
Tested diff errors on both dark and light modes.

Reviewed at https://reviews.reviewboard.org/r/13879/