Fix the default background color for review request infobox summaries.
The recent work on the review request infobox missed updating the
background color for the summary area when there weren't any issues or
Ship Its. This was showing up as white on bright grey.

We now use a much more appropriate background color, a dark grey, a
shade or two lighter than the background of the infobox itself. This is
far more readable and better fits the UI.

Testing Done:
Tested the infobox in light and dark modes, with and without Ship Its and
open issues.

Reviewed at https://reviews.reviewboard.org/r/13880/