Convert RB.DraftReviewRequest to TypeScript/spina
This change converts the DraftReviewRequest resource model to
TypeScript.

The one complication in here is a change in behavior with mixins
where the `url()` implementation in `DraftResourceModelMixin` was
taking priority over the one in `DraftReviewRequest`. In the vast
majority of draft resource cases, the implementation in the mixin
is what we want, but review requests are special.

Testing Done:
- Ran js-tests.
- Ran setup.py build_media.

Reviewed at https://reviews.reviewboard.org/r/14086/