Fix auth page description text color in dark mode.
Authentication pages were hard-coding a color for paragraph text in the
auth header. This wasn't noticible on most of the auth pages, because
when people are logged out, we show the default light theme. For
web-based RBTools authentication, we do have pages that the user could
see in dark mode. This change fixes it to use the `--ink-p-fg-weak`
color instead.

Testing Done:
Was able to see description text for the web-based login pages.

Reviewed at https://reviews.reviewboard.org/r/13894/