Add dark mode and modern tabs to the text and Markdown review UI.
The text review UI has been updated to support dark mode styling. The
text content already supported dark mode, but the custom tabs and the
overridden styles for the filename header did not.

We now use `rb-c-tabs` for the tabs, helping that fit in with much of
the UI. We also use the default diff viewer header styling that we
inherit for the filename information.

Tabs had to be updated to ensure we correctly override the `color` and
`text-decoration` for the label, in case `<a>` is used (which is the
case here).

With this, we now have full dark mode support for text and Markdown
review.

Testing Done:
Tested in light and dark modes with single revisions, diffs, plain
text, and Markdown files.

Reviewed at https://reviews.reviewboard.org/r/13890/